restructure of cpp folder, added header and cc files, now bruteforcesolution class has to be implemented and tested
