+++ knapsack/cpp/Item.cc
@@ -1,0 +1,36 @@
+#include "Item.h"
+
+Item::Item(long w, long v) {
+    add_weight(w);
+    add_value(v);
+}
+
+void Item::add_weight(long w) {
+    if (w > 0) {
+        weight = w;
+    } else {
+        cout << "Weight cannot be a negative number!\nSetting weight to 0." << endl;
+        weight = 0;
+    }
+}
+
+void Item::add_value(long v) { 
+    if (v > 0) {
+        value = v;
+    } else {
+        cout << "Value cannot be zero or a negative number!\nSetting value to 0." << endl;
+        value = 0;
+    }
+}
+
+long Item::get_weight() {
+    return weight;
+}
+
+long Item::get_value() {
+    return value;
+}
+
+string Item::state() {
+    return "Weight: " + to_string(weight) + "\nValue: " + to_string(value);
+}
