+++ knapsack/cpp/Solution.h
@@ -1,0 +1,10 @@
+#include <string>
+#include <iostream>
+#include "Item.h"
+
+using std::cout, std::endl, std::vector, std::to_string, std::equal;
+
+void test_Item();
+void test_item_weight(Item item, long w);
+void test_item_value(Item item, long v);
+void test_item_state(Item item, long w, long v);
